Move instance-related functions to new `instance.go` file
