+++ pkg/sandboxesapi/config.go
@@ -1,9 +1,28 @@
 package sandboxesapi
 
 import (
+	"fmt"
+
 	"github.com/keboola/go-client/pkg/client"
 	"github.com/keboola/go-client/pkg/storageapi"
 )
+
+func GetSandboxID(c *storageapi.Config) (SandboxID, error) {
+	id, found, err := c.Content.GetNested("parameters.id")
+	if err != nil {
+		return "", err
+	}
+	if !found {
+		return "", fmt.Errorf("config is missing sandboxId")
+	}
+
+	out, ok := id.(string)
+	if !ok {
+		return "", fmt.Errorf("config.parameters.id is not a string")
+	}
+
+	return SandboxID(out), nil
+}
 
 func GetConfigRequest(branchId BranchID, configId ConfigID) client.APIRequest[*storageapi.Config] {
 	key := storageapi.ConfigKey{
