+++ pkg/sandboxesapi/sandbox.go
@@ -13,37 +13,6 @@ import (
 type BranchID = storageapi.BranchID
 type ConfigID = storageapi.ConfigID
 
-type SandboxID string
-
-func (v SandboxID) String() string {
-	return string(v)
-}
-
-type Sandbox struct {
-	ID       SandboxID `json:"id"`
-	Type     string    `json:"type"`
-	Size     string    `json:"size"` // Only exists for container sandboxes (Python, R)
-	Active   bool      `json:"active"`
-	Shared   bool      `json:"shared"`
-	User     string    `json:"user"`
-	Host     string    `json:"host"`
-	Url      string    `json:"url"`
-	Password string    `json:"password"`
-	Created  Time      `json:"createdTimestamp"`
-	Updated  Time      `json:"updatedTimestamp"`
-	Start    Time      `json:"startTimestamp"`
-	// Workspace details - only exists for Snowflake sandboxes
-	Details *Details `json:"workspaceDetails"`
-}
-
-type Details struct {
-	Connection struct {
-		Database  string `json:"database"`
-		Schema    string `json:"schema"`
-		Warehouse string `json:"warehouse"`
-	} `json:"connection"`
-}
-
 type SandboxWithConfig struct {
 	Sandbox *Sandbox
 	Config  *storageapi.Config
@@ -104,23 +73,6 @@ func SupportsSizes(typ string) bool {
 	default:
 		return false
 	}
-}
-
-func GetSandboxID(c *storageapi.Config) (SandboxID, error) {
-	id, found, err := c.Content.GetNested("parameters.id")
-	if err != nil {
-		return "", err
-	}
-	if !found {
-		return "", fmt.Errorf("config is missing sandboxId")
-	}
-
-	out, ok := id.(string)
-	if !ok {
-		return "", fmt.Errorf("config.parameters.id is not a string")
-	}
-
-	return SandboxID(out), nil
 }
 
 func Create(
@@ -270,22 +222,4 @@ func List(
 		})
 	}
 	return out, nil
-
-}
-
-func GetInstanceRequest(sandboxId SandboxID) client.APIRequest[*Sandbox] {
-	result := &Sandbox{}
-	request := newRequest().
-		WithResult(&result).
-		WithGet("sandboxes/{sandboxId}").
-		AndPathParam("sandboxId", sandboxId.String())
-	return client.NewAPIRequest(result, request)
-}
-
-func ListInstancesRequest() client.APIRequest[*[]*Sandbox] {
-	result := make([]*Sandbox, 0)
-	request := newRequest().
-		WithResult(&result).
-		WithGet("sandboxes")
-	return client.NewAPIRequest(&result, request)
 }
